+++ resources/views/admin/data_pencalon.blade.php
@@ -1,10 +1,30 @@
 @extends('aplikasi.layout')
+
+@section('admin')
+<li class="sub-menu active" >
+<a href="javascript:;" class="">
+                          <i class="icon_desktop"></i>
+                          <span>FITUR Admin</span>
+                          <span class="menu-arrow arrow_carrot-right"></span>
+                      </a>
+                      <ul class="sub">
+                          <li><a class="" href="/data-kategori">data kategori</a></li>
+                          @yield('sidebar')
+                          <li><a class="" href="/data-kelas">data kelas</a></li>
+                          <li><a class="" href="/data-user">data user</a></li>
+                          <li><a class="" href="/data-pencalon">data pencalonan</a></li>
+                          <li><a class="" href="/data-calon">data kandidat</a></li>
+
+                      </ul>
+                  </li>
+@endsection
+
 @section('content')
 <div class="row">
   <div class="col-md-12">
     <section class="panel">
       <header class="panel-heading">
-        <h3>Data Pencalon</h3>
+        <h3>{{Request::path()==='data-calon'?'Data kandidat' : 'Data pencalon'}}</h3>
       </header>
       <div class="panel-body">
         <table class="table table-striped table-advance table-hover">
@@ -42,8 +62,9 @@
                 @if($data->status == 1)
                   <a class="btn btn-primary" href="/terima-calon/{{$data->id}}"><i class="icon_plus_alt "></i></a>
                 @endif
-
+                 @if(Request::path()=='data-pencalon')
                   <a class="btn btn-danger" href="/pencalon/{{$data->id}}/hapus"><i class="icon_trash"></i></a>
+                  @endif
                 </div>
               </td>
             </tr>
